+++ tests/EnvTest.php
@@ -9,4 +9,10 @@ class EnvTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(123, \JmesPath\Env::search('foo', $data));
         $this->assertEquals(123, \JmesPath\Env::search('foo', $data));
     }
+
+    public function testSearchesWithFunction()
+    {
+        $data = array('foo' => 123);
+        $this->assertEquals(123, \JmesPath\search('foo', $data));
+    }
 }